Use redcarpet directly instead of going through github-markup.
github-markup supports multiple markups. For markdown it calls redcarpet with default arguments, that is, without enabling any markdown extensions. We do need to enable the fenced code blocks (blocks delimited with 3 backticks) but we don't need the support for any other markup flavors.
This commit removes the github-markup dependency and calls redcarpet directly with fenced code support enabled.
Code blocks are supposed to be marked with fences so I disabled indented codeblocks as a safety measure.